refactored mainMenu into MainMenu Interface and screens implementing mm
