+++ view/MainMenu/MainMenuScreen.py
@@ -7,11 +7,10 @@ from view.BasicView import BasicView
 from controller.ControllerView import ControllerMainMenu
 
 
+class MainMenuScreen(BasicView):
 
-class MainMenu(BasicView):
-
-    def __init__(self, window: pygame.display, controller: ControllerMainMenu) -> None:
-        super(MainMenu, self).__init__(window, controller)
+    def __init__(self, window: pygame.display, controller: ControllerMainMenu):
+        super(MainMenuScreen, self).__init__(window, controller)
 
         self.window_width, self.window_height = pygame.display.get_surface().get_size()
 
@@ -24,17 +23,15 @@ class MainMenu(BasicView):
                                       (self.window_width / 4, self.window_height / 2)),
             manager=self.manager)
 
-
         self.__padding = self.container.rect.width / 10
         self.__buttonSize = (self.container.rect.width / 3, self.container.rect.width / 12)
 
         self.background = pygame.Surface((self.window_width, self.window_height))
         self.background.fill(self.manager.ui_theme.get_colour(None, None, 'dark_bg'))
 
-
         self.start_game_button = pygame_gui.elements.UIButton(
             relative_rect=pygame.Rect((self.container.rect.centerx,
-                                       self.container.rect.centery + self.__padding*len(self.container.elements)),
+                                       self.container.rect.centery + self.__padding * len(self.container.elements)),
                                       self.__buttonSize),
             text="Start Game",
             manager=self.manager,
@@ -42,7 +39,7 @@ class MainMenu(BasicView):
         )
         self.help_button = pygame_gui.elements.UIButton(
             relative_rect=pygame.Rect((self.container.rect.centerx,
-                                       self.container.rect.centery + self.__padding*len(self.container.elements)),
+                                       self.container.rect.centery + self.__padding * len(self.container.elements)),
                                       self.__buttonSize),
             text="Help",
             manager=self.manager,
@@ -50,7 +47,7 @@ class MainMenu(BasicView):
         )
         self.settings_button = pygame_gui.elements.UIButton(
             relative_rect=pygame.Rect((self.container.rect.centerx,
-                                       self.container.rect.centery + self.__padding*len(self.container.elements)),
+                                       self.container.rect.centery + self.__padding * len(self.container.elements)),
                                       self.__buttonSize),
             text="Settings",
             manager=self.manager,
@@ -58,22 +55,16 @@ class MainMenu(BasicView):
         )
         self.end_game_button = pygame_gui.elements.UIButton(
             relative_rect=pygame.Rect((self.container.rect.centerx,
-                                       self.container.rect.centery + self.__padding*len(self.container.elements)),
+                                       self.container.rect.centery + self.__padding * len(self.container.elements)),
                                       self.__buttonSize),
             text="End Game",
             manager=self.manager,
             container=self.container
         )
 
-        #load title image
+        # load title image
         self.titleImage = pygame.image.load("assets/Menu/TitleImage.png")
-        logging.info("MainMenu init done")
-
-    def init(self) -> None:
-        """
-        Not yet implemented
-        :return:    None
-        """
+        logging.info(f"MainMenuScreen init done")
 
     def draw(self) -> None:
         pygame.draw.circle(self.window, (255, 0, 0), (250, 250), 50)
@@ -81,7 +72,7 @@ class MainMenu(BasicView):
 
         self.window.blit(self.background, (0, 0))
         self.window.blit(self.titleImage,
-                         (self.window_width/2 - self.titleImage.get_rect().width/2, self.window_height * .25))
+                         (self.window_width / 2 - self.titleImage.get_rect().width / 2, self.window_height * .25))
         self.manager.draw_ui(self.window)
 
         pygame.display.update()
